Redesign front-end and add script
Redesigning front end using more simple/brutal layout. Using GSAP
to do some basic animations. Added a script to delete the
ipTokenCount entries. This script should be cron'd on the server
to allow users to post after reaching the MAX_DAILY_TOKENS limit
as defined in restrict.js.

This commit is still a WIP as far as the front end goes.
index_original is just a backup of the original layout.